+++ plom/db/buildPlomDB.py
@@ -8,7 +8,8 @@ import logging
 import random
 
 from plom import check_version_map, make_random_version_map
-from plom.db import PlomDB
+
+# from plom.db import PlomDB
 
 
 log = logging.getLogger("DB")
@@ -70,8 +71,8 @@ def buildSpecialRubrics(spec, db):
                 )
 
 
-def buildExamDatabaseFromSpec(spec, db, version_map=None):
-    """Build metadata for exams from spec and insert into the database.
+def initialiseExamDatabaseFromSpec(spec, db, version_map=None):
+    """Build metadata for exams from spec but do not build tests in DB.
 
     Arguments:
         spec (dict): exam specification, see :func:`plom.SpecVerifier`.
@@ -82,18 +83,19 @@ def buildExamDatabaseFromSpec(spec, db, version_map=None):
             see :func:`plom.finish.make_random_version_map`.
 
     Returns:
-        bool: True if succuess.
-        str: a status string, one line per test, ending with an error if failure.
+        (dict): the question-version map
 
     Raises:
-        ValueError: if database already populated.
+        ValueError: if database already populated, or attempt to
+        build paper n without paper n-1.
         KeyError: invalid question selection scheme in spec.
     """
     if db.is_paper_database_populated():
         raise ValueError("Database already populated")
 
-    # TODO: perhaps this should be called separately...
     buildSpecialRubrics(spec, db)
+    if not db.createReplacementBundle():
+        raise ValueError("Could not create bundle for replacement pages")
 
     if not version_map:
         # TODO: move reproducible random seed support to the make fcn?
@@ -101,66 +103,4 @@ def buildExamDatabaseFromSpec(spec, db, version_map=None):
         version_map = make_random_version_map(spec)
     check_version_map(version_map)
 
-    ok = True
-    status = ""
-    # build bundles for annotation images
-    # for q in range(1, 1 + spec["numberOfQuestions"]):
-    # for v in range(1, 1 + spec["numberOfVersions"]):
-    # pass
-    # if not db.createAnnotationBundle(q, v):
-    #     ok = False
-    #     status += "Error making bundle for q.v={}.{}".format(q, v)
-    # build bundle for replacement pages (for page-not-submitted images)
-
-    if not db.createReplacementBundle():
-        ok = False
-        status += "Error making bundle for replacement pages"
-
-    for t in range(1, spec["numberToProduce"] + 1):
-        log.info(
-            "Creating DB entry for test {} of {}.".format(t, spec["numberToProduce"])
-        )
-        if db.createTest(t):
-            status += "DB entry for test {:04}:".format(t)
-        else:
-            status += " Error creating"
-            ok = False
-
-        if db.createIDGroup(t, [spec["idPage"]]):
-            status += " ID"
-        else:
-            status += " Error creating idgroup"
-            ok = False
-
-        if db.createDNMGroup(t, spec["doNotMarkPages"]):
-            status += " DNM"
-        else:
-            status += "Error creating DoNotMark-group"
-            ok = False
-
-        for g in range(spec["numberOfQuestions"]):  # runs from 0,1,2,...
-            gs = str(g + 1)  # now a str and 1,2,3,...
-            v = version_map[t][g + 1]
-            assert v in range(1, spec["numberOfVersions"] + 1)
-            if spec["question"][gs]["select"] == "fix":
-                vstr = "f{}".format(v)
-                assert v == 1
-            elif spec["question"][gs]["select"] == "shuffle":
-                vstr = "v{}".format(v)
-            # elif spec["question"][gs]["select"] == "param":
-            #     vstr = "p{}".format(v)
-            else:
-                raise KeyError(
-                    'Invalid spec: question {} "select" of "{}" is unexpected'.format(
-                        gs, spec["question"][gs]["select"]
-                    )
-                )
-            # if db.createQGroup(t, int(gs), v, spec["question"][gs]["pages"], int(spec["question"][gs]["mark"])):
-            if db.createQGroup(t, int(gs), v, spec["question"][gs]["pages"]):
-                status += " Q{}{}".format(gs, vstr)
-            else:
-                status += "Error creating Question {} ver {}".format(gs, vstr)
-                ok = False
-        status += "\n"
-
-    return ok, status
+    return version_map