+++ plom/plom_exceptions.py
@@ -63,6 +63,10 @@ class PlomBadTagError(PlomBenignException):
     pass
 
 
+class PlomDatabaseCreationError(PlomBenignException):
+    pass
+
+
 class PlomExistingDatabase(PlomBenignException):
     """The database has already been populated."""